Core: Move InMemoryCatalog from test to core
