+++ core/src/main/java/org/apache/iceberg/inmemory/InMemoryInputFile.java
@@ -16,11 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg.io;
+package org.apache.iceberg.inmemory;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.util.UUID;
+import org.apache.iceberg.io.InputFile;
+import org.apache.iceberg.io.SeekableInputStream;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 
 public class InMemoryInputFile implements InputFile {
