+++ core/src/test/java/org/apache/iceberg/inmemory/TestInMemoryOutputFile.java
@@ -16,23 +16,24 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg.io;
+package org.apache.iceberg.inmemory;
 
 import java.io.IOException;
-import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import org.assertj.core.api.Assertions;
-import org.junit.Assert;
 import org.junit.Test;
 
-public class TestInMemoryInputFile {
+public class TestInMemoryOutputFile {
   @Test
-  public void testReadAfterClose() throws IOException {
-    InMemoryInputFile inputFile =
-        new InMemoryInputFile("abc".getBytes(StandardCharsets.ISO_8859_1));
-    InputStream inputStream = inputFile.newStream();
-    Assert.assertEquals('a', inputStream.read());
-    inputStream.close();
-    Assertions.assertThatThrownBy(inputStream::read).hasMessage("Stream is closed");
+  public void testWriteAfterClose() throws IOException {
+    InMemoryOutputFile outputFile = new InMemoryOutputFile();
+    OutputStream outputStream = outputFile.create();
+    outputStream.write('a');
+    outputStream.write('b');
+    outputStream.close();
+    Assertions.assertThatThrownBy(() -> outputStream.write('c')).hasMessage("Stream is closed");
+    Assertions.assertThat(outputFile.toByteArray())
+        .isEqualTo("ab".getBytes(StandardCharsets.ISO_8859_1));
   }
 }
